Introduce createDeps
