+++ models/recruit/src/index.ts
@@ -14,7 +14,7 @@
 //
 
 import type { Employee } from '@anticrm/contact'
-import { Class, Data, Doc, FindOptions, Ref, Space, Timestamp } from '@anticrm/core'
+import { Doc, FindOptions, Ref, Timestamp } from '@anticrm/core'
 import { Builder, Collection, Model, Prop, TypeBoolean, TypeDate, TypeRef, TypeString, UX } from '@anticrm/model'
 import attachment from '@anticrm/model-attachment'
 import chunter from '@anticrm/model-chunter'
@@ -25,7 +25,6 @@ import view from '@anticrm/model-view'
 import workbench from '@anticrm/model-workbench'
 import type { IntlString } from '@anticrm/platform'
 import { Applicant, Candidate, Candidates, Vacancy } from '@anticrm/recruit'
-import { createDefaultKanbanTemplate } from '@anticrm/task'
 import recruit from './plugin'
 
 @Model(recruit.class.Vacancy, task.class.SpaceWithStates)
@@ -277,11 +276,6 @@ export function createModel (builder: Builder): void {
     action: task.action.CreateTask
   })
 
-  builder.createDoc(task.class.Sequence, task.space.Sequence, {
-    attachedTo: recruit.class.Applicant,
-    sequence: 0
-  })
-
   builder.createDoc(
     task.class.KanbanTemplateSpace,
     core.space.Model,
@@ -295,24 +289,8 @@ export function createModel (builder: Builder): void {
     },
     recruit.space.VacancyTemplates
   )
-
-  // eslint-disable-next-line @typescript-eslint/no-floating-promises
-  createDefaultKanbanTemplate(async <T extends Doc>(
-    props: {
-      id?: Ref<T>
-      space: Ref<Space>
-      class: Ref<Class<T>>
-    },
-    attrs: Data<T>
-  ): Promise<void> => {
-    builder.createDoc(
-      props.class,
-      props.space,
-      attrs,
-      props.id
-    )
-  })
 }
 
-export { recruitOperation } from './migration'
 export { default } from './plugin'
+export { recruitOperation } from './migration'
+export { createDeps } from './creation'
