+++ models/recruit/src/migration.ts
@@ -13,61 +13,15 @@
 // limitations under the License.
 //
 
-import core, { TxOperations } from '@anticrm/core'
 import { MigrateOperation, MigrationClient, MigrationResult, MigrationUpgradeClient } from '@anticrm/model'
-import contact, { DOMAIN_CONTACT } from '@anticrm/model-contact'
-import { DOMAIN_TASK } from '@anticrm/model-task'
-import { createDefaultKanbanTemplate } from '@anticrm/recruit'
-import recruit from './plugin'
 
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
 function logInfo (msg: string, result: MigrationResult): void {
   if (result.updated > 0) {
     console.log(`Recruit: Migrate ${msg} ${result.updated}`)
   }
 }
 export const recruitOperation: MigrateOperation = {
-  async migrate (client: MigrationClient): Promise<void> {
-    logInfo(
-      'done for Applicants',
-      await client.update(
-        DOMAIN_TASK,
-        { _class: recruit.class.Applicant, doneState: { $exists: false } },
-        { doneState: null }
-      )
-    )
-
-    logInfo(
-      '$move employee => assignee',
-      await client.update(
-        DOMAIN_TASK,
-        { _class: recruit.class.Applicant, employee: { $exists: true } },
-        { $rename: { employee: 'assignee' } }
-      )
-    )
-
-    const employees = (await client.find(DOMAIN_CONTACT, { _class: contact.class.Employee })).map((emp) => emp._id)
-
-    // update assignee to unassigned if there is no employee exists.
-    logInfo(
-      'applicants wrong assignee',
-      await client.update(
-        DOMAIN_TASK,
-        { _class: recruit.class.Applicant, assignee: { $not: { $in: employees } } },
-        { assignee: null }
-      )
-    )
-  },
-  async upgrade (client: MigrationUpgradeClient): Promise<void> {
-    console.log('Recruit: Performing model upgrades')
-
-    const ops = new TxOperations(client, core.account.System)
-    if (await client.findOne(core.class.TxCreateDoc, { objectId: recruit.template.DefaultVacancy }) === undefined) {
-      await createDefaultKanbanTemplate(async (
-        props,
-        attrs
-      ): Promise<void> => {
-        await ops.createDoc(props.class, props.space, attrs, props.id)
-      })
-    }
-  }
+  async migrate (client: MigrationClient): Promise<void> {},
+  async upgrade (client: MigrationUpgradeClient): Promise<void> {}
 }
