+++ plugins/recruit/src/index.ts
@@ -17,7 +17,7 @@ import type { Person } from '@anticrm/contact'
 import type { Class, Ref, Space, Timestamp } from '@anticrm/core'
 import type { Asset, Plugin } from '@anticrm/platform'
 import { plugin } from '@anticrm/platform'
-import { CreateFn, createKanbanTemplate, KanbanTemplate, KanbanTemplateSpace, SpaceWithStates, Task } from '@anticrm/task'
+import type { KanbanTemplateSpace, SpaceWithStates, Task } from '@anticrm/task'
 
 /**
  * @public
@@ -79,36 +79,7 @@ const recruit = plugin(recruitId, {
   },
   space: {
     VacancyTemplates: '' as Ref<KanbanTemplateSpace>
-  },
-  template: {
-    DefaultVacancy: '' as Ref<KanbanTemplate>
   }
 })
 
 export default recruit
-
-const defaultKanban = {
-  states: [
-    { color: '#7C6FCD', title: 'HR Interview' },
-    { color: '#6F7BC5', title: 'Technical Interview' },
-    { color: '#77C07B', title: 'Test task' },
-    { color: '#A5D179', title: 'Offer' }
-  ],
-  doneStates: [
-    { isWon: true, title: 'Won' },
-    { isWon: false, title: 'Lost' }
-  ]
-}
-
-/**
- * @public
- */
-export const createDefaultKanbanTemplate = async (create: CreateFn): Promise<void> => {
-  await createKanbanTemplate(create)({
-    kanbanId: recruit.template.DefaultVacancy,
-    space: recruit.space.VacancyTemplates,
-    title: 'Default vacancy',
-    states: defaultKanban.states,
-    doneStates: defaultKanban.doneStates
-  })
-}
